Moved ProgressRequestBody from api to internal package
