+++ lib/src/main/java/com/ibm/mobilefirstplatform/clientsdk/android/core/internal/ProgressRequestBody.java
@@ -1,6 +1,7 @@
-package com.ibm.mobilefirstplatform.clientsdk.android.core.api;
+package com.ibm.mobilefirstplatform.clientsdk.android.core.internal;
 
 
+import com.ibm.mobilefirstplatform.clientsdk.android.core.api.ProgressListener;
 import com.ibm.mobilefirstplatform.clientsdk.android.logger.api.Logger;
 import com.squareup.okhttp.MediaType;
 import com.squareup.okhttp.RequestBody;
